auth-service/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
cryptography==42.0.8
Werkzeug==3.1.3
gunicorn==21.2.0
requests==2.32.3
psycopg2-binary==2.9.9
alembic==1.13.2
SQLAlchemy==2.0.36
